+++ lib/openapi_first/utils.rb
@@ -1,13 +1,7 @@
 # frozen_string_literal: true
 
-require 'deep_merge/core'
-
 module OpenapiFirst
   module Utils
-    def self.deep_merge!(dest, source)
-      DeepMerge.deep_merge!(source, dest)
-    end
-
     class StringKeyedHash
       def initialize(original)
         @orig = original